feat: upgrade code and added middleware valid

+++ src/app.js
@@ -1,12 +1,12 @@
 require("dotenv").config();
 const express = require("express");
+const app = express();
+
 const jwt = require("jsonwebtoken");
 const pool = require("./config/db");
-const app = express();
+const auth = require("./auth/auth");
 const validPost = require("./valid/post");
 const validUser = require("./valid/user");
-const auth = require("./auth/auth")
-
 app.use(express.json());
 
 const formattedDataLocal = (data) => {
@@ -26,7 +26,6 @@ const resFormatted = (dataRows) => {
 app.get("/", (req, res) => {
   res.send("<h1>Olá redes sociaigs</h1>");
 });
-
 app.get("/user", async (req, res) => {
   try {
     const getUserData = await pool.query(`
@@ -37,7 +36,6 @@ app.get("/user", async (req, res) => {
     res.status(500).json({ err: "Não foi possivel carregar as informações" });
   }
 });
-
 app.get("/posts", async (req, res) => {
   try {
     const getData = await pool.query(`
@@ -57,7 +55,6 @@ app.get("/posts", async (req, res) => {
     res.status(500).json({ erro: "FALHA AO FAZER A REQUISIÇÃO" });
   }
 });
-
 app.post("/user", validUser, async (req, res) => {
   try {
     const { name, email, password } = req.body;
@@ -77,10 +74,9 @@ app.post("/user", validUser, async (req, res) => {
     res.status(500).json({ err: "Não foi possivel criar o post" });
   }
 });
-
 app.post("/post", auth, validPost, async (req, res) => {
   try {
-    const {title, content } = req.body;
+    const { title, content } = req.body;
     const sendData = await pool.query(
       `
             INSERT INTO tb_post(user_id, title, content)
@@ -101,7 +97,7 @@ app.post("/login", async (req, res) => {
     const { email, password } = req.body;
     const user = await pool.query(
       `
-        SELECT * FROM tb_user WHERE email= $1
+        SELECT * FROM tb_user WHERE email=$1
     `,
       [email],
     );
@@ -116,18 +112,27 @@ app.post("/login", async (req, res) => {
     const token = jwt.sign({ id: user.rows[0].id }, process.env.JWT_SECRET, {
       expiresIn: "1h",
     });
-    console.log("login secreto", token);
-
     res.status(200).json({ token });
   } catch (err) {
     console.log(err);
   }
 });
-
-app.put("/post/:id", validPost, async (req, res) => {
+app.put("/post/:id", auth, validPost, async (req, res) => {
   try {
     const { id } = req.params;
+    const idPost = req.user.id;
     const { title, content } = req.body;
+
+    const post = await pool.query(`SELECT * FROM tb_post WHERE id=$1`, [id]);
+
+    if (post.rows.length === 0) {
+      return res.status(404).json({ mgs: "Post não encontrado" });
+    }
+
+    if (post.user_id === req.user.id) {
+      return res.status(403).json({ msg: "Sem permissão" });
+    }
+
     const updatedValue = await pool.query(
       `
         UPDATE tb_post SET title=$1, content=$2 
@@ -135,6 +140,8 @@ app.put("/post/:id", validPost, async (req, res) => {
         `,
       [title, content, id],
     );
+
+    console.log(updatedValue.rows)
 
     res.json({
       mensagem: "Post atualizado com sucesso",
@@ -144,10 +151,18 @@ app.put("/post/:id", validPost, async (req, res) => {
     res.status(500).json({ err: "Não foi possivel atualizar as informações" });
   }
 });
-
-app.delete("/posts/:id", async (req, res) => {
+app.delete("/posts/:id", auth, async (req, res) => {
   try {
     const { id } = req.params;
+    const post = await pool.query(`SELECT * FROM tb_post WHERE id=$1`, [id]);
+
+    if (post.rows.length === 0) {
+      return res.status(404).json({ mgs: "Post não encontrado" });
+    }
+
+    if (post.user_id === req.user.id) {
+      return res.status(403).json({ msg: "Sem permissão" });
+    }
     const deleteData = await pool.query(
       `DELETE FROM tb_post WHERE id=$1 RETURNING *`,
       [id],
@@ -160,8 +175,7 @@ app.delete("/posts/:id", async (req, res) => {
     res.status(500).json({ err: "Não foi possivel deletar as informações" });
   }
 });
-
-app.delete("/user/:id", async (req, res) => {
+app.delete("/user/:id", auth, async (req, res) => {
   try {
     const { id } = req.params;
     const deleteData = await pool.query(
